% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_angle.R
\name{car_to_deg}
\alias{car_to_deg}
\title{Cartesian x, y to angle degrees}
\usage{
car_to_deg(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates in the Cartesian plane}

\item{y}{a numeric vector of y coordinates in the Cartesian plane}
}
\value{
a numeric vector of corresponding angle degrees (anticlockwise direction)
}
\description{
Convert Cartesian coordinates x and y to angle degrees
}
\details{
By approaching the problem with the use of complex numbers (strange enough but it's easier!)
If z = x + i*y with real x and y
r = Mod(z) = sqrt(x^2 + y^2)
phi = Arg(z)
x = r * cos(phi)
y = r * sin(phi)
}
\examples{
x <- c(1, 1, 0, -1, -1, -1,  0,  1)
y <- c(0, 1, 1,  1,  0, -1, -1, -1)
car_to_deg(x, y)
car_to_deg(1, 1)   # Expected: 45 degrees
car_to_deg(0, 1)   # Expected: 90 degrees
car_to_deg(-1, 1)  # Expected: 135 degrees
car_to_deg(-1, 0)  # Expected: 180 degrees
car_to_deg(-1, -1) # Expected: 225 degrees
car_to_deg(0, -1)  # Expected: 270 degrees
car_to_deg(1, -1)  # Expected: 315 degrees
car_to_deg(1, 0)   # Expected: 0 degrees
}
