% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_angle.R
\name{car_to_deg_N}
\alias{car_to_deg_N}
\title{Cartesian x, y to angle degrees from the North (compass direction)}
\usage{
car_to_deg_N(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates in the Cartesian plane}

\item{y}{a numeric vector of y coordinates in the Cartesian plane}
}
\value{
a numeric vector of corresponding angle degrees direction from the North (compass direction)
}
\description{
Convert Cartesian coordinates x and y to angle degrees from the North direction, i.e. the compass direction
}
\details{
By approaching the problem with the use of complex numbers (strange enough but it's easier!)
If z = x + i*y with real x and y
r = Mod(z) = sqrt(x^2 + y^2)
phi = Arg(z)
x = r * cos(phi)
y = r * sin(phi)
}
\examples{
x <- c(1, 1, 0, -1, -1, -1,  0,  1)
y <- c(0, 1, 1,  1,  0, -1, -1, -1)
car_to_deg_N(x, y)
car_to_deg_N(1, 1)   # Expected: 315 degrees (North = 0, East = 90)
car_to_deg_N(0, 1)   # Expected: 0 degrees (North)
car_to_deg_N(-1, 1)  # Expected: 45 degrees (North-East)
car_to_deg_N(-1, 0)  # Expected: 90 degrees (East)
car_to_deg_N(-1, -1) # Expected: 135 degrees (South-East)
car_to_deg_N(0, -1)  # Expected: 180 degrees (South)
car_to_deg_N(1, -1)  # Expected: 225 degrees (South-West)
car_to_deg_N(1, 0)   # Expected: 270 degrees (West)
}
