% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_conc.R
\name{conc_from_to}
\alias{conc_from_to}
\title{Convert concentrations v/v or mass/v}
\usage{
conc_from_to(
  value,
  mw,
  temp = 293.15,
  press = 1013,
  from = c("ppm", "ppb", "mgm3", "ugm3"),
  to = c("ppm", "ppb", "mgm3", "ugm3")
)
}
\arguments{
\item{value}{numeric vector, concentration value to be converted from to}

\item{mw}{numeric vector, molecular weight of chemical substance to be converted}

\item{temp}{numeric vector, temperature}

\item{press}{numeric vector, pressure}

\item{from}{character, from measure unit}

\item{to}{character, to measure unit}
}
\value{
vector of converted concentrations
}
\description{
Convert concentrations volume/volume or mass/volume from one unit to another.
Input and output measure units are: ppm, ppb, mg/m3, ug/m3, in any sensible (!) direction.
}
\examples{
conc_from_to(10, 44.01, from = "ppm", to = "mgm3")
conc_from_to(1000, 44.01, from = "ppb", to = "mgm3")
conc_from_to(1, 44.01, from = "mgm3", to = "ppm")
conc_from_to(1000, 44.01, from = "ugm3", to = "ppm")

}
