% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord.R
\name{coord_convert}
\alias{coord_convert}
\title{Convert spatial coordinates}
\usage{
coord_convert(x, y, epsg_in = 32632, epsg_out = 4326)
}
\arguments{
\item{x}{vector of x/lon coordinates}

\item{y}{vector of y/lat coordinates}

\item{epsg_in}{number corresponding to EPSG code of initial CRS, default to UTM32, no error check implemented}

\item{epsg_out}{number corresponding to EPSG code of converted CRS, default to WGS84, no error check implemented}
}
\value{
a dataframe of converted x/lon and y/lat coordinates with indication of final EPSG code
}
\description{
Conversion from one CRS to another by setting the EPSG codes
}
\examples{
coord_convert(738239.9, 5073656)
coord_convert(738239.9, 5073656, epsg_in = 32632, epsg_out = 32633)

}
