% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_yday.R
\name{date_to_yday}
\alias{date_to_yday}
\title{From date to year-day}
\usage{
date_to_yday(mydate)
}
\arguments{
\item{mydate}{string in the form "YYY-MM-DD"representing input date}
}
\value{
string in the form "YYYY-YDAY" representing output year-day (a.k.a. julian-day)
}
\description{
Convert date in the form "YYY-MM-DD" to year-day (a.k.a. julian-day) in the form "YYYY-YDAY"
}
\examples{
date_to_yday("2022-05-30")
}
