% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_dd.R
\name{dd_to_dms}
\alias{dd_to_dms}
\title{From decimal degrees to degrees, minutes, seconds}
\usage{
dd_to_dms(dd)
}
\arguments{
\item{dd}{numeric vector of decimal degrees}
}
\value{
character vector of degrees minute seconds
}
\description{
Convert degrees minutes seconds to decimal degrees
}
\examples{
dd_to_dms(30.25556)      # Expected: "30d 15m 20.00s"
}
