% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_angle.R
\name{deg_N_to_car}
\alias{deg_N_to_car}
\title{From angles degrees of North (compass direction) to Cartesian x, y}
\usage{
deg_N_to_car(deg, r = 1)
}
\arguments{
\item{deg}{vector of degrees North}

\item{r}{the radius or magnitude of the vector in the Cartesian coordinate system.}
}
\value{
a tibble with degrees, phi in radians, x and y
}
\description{
Convert from the North direction in angles degrees (i.e. the compass direction) to the cartesian coordinates x and y
#' What is r?
It defines how far the point is from the origin (0,0)
If r = 1 (default), the function converts only the direction (angle) into Cartesian coordinates.
If r > 1, it scales the coordinates proportionally.
If r < 1, it shrinks the coordinates proportionally.
If r = 0, it always returns (0,0), regardless of deg.
}
\examples{
deg_N_to_car(0)     # Expected: (0, 1) -> North
deg_N_to_car(90)    # Expected: (1, 0) -> East
deg_N_to_car(180)   # Expected: (0, -1) -> South
deg_N_to_car(270)   # Expected: (-1, 0) -> West
deg_N_to_car(45, 2) # Expected: (√2, √2) -> North-East
}
