% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_dd.R
\name{dms_to_dd}
\alias{dms_to_dd}
\title{From degrees, minutes, seconds to decimal degrees}
\usage{
dms_to_dd(s)
}
\arguments{
\item{s}{character vector degrees, minutes, seconds}
}
\value{
numeric vector of decimal degrees
}
\description{
Convert degrees minutes seconds to decimal degrees
}
\examples{
dms_to_dd("30 15 20") # Expected: 30.25556
}
