% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ppm.R
\name{mgm3_to_ppm}
\alias{mgm3_to_ppm}
\title{mg/m3 to ppm}
\usage{
mgm3_to_ppm(mgm3, mw, temp = 293.15, press = 1013)
}
\arguments{
\item{mgm3}{numeric vector of mass concentrations in mg per cubic meter (m^-3)}

\item{mw}{molecular weight (g/mol)}

\item{temp}{temperature in K (default = 293 K)}

\item{press}{pressure in hPa (default = 1013 hPa)}
}
\value{
numeric vector of volume concentrations in parts per billion (10^-6)
}
\description{
convert from mg/m3 to ppm
}
\examples{
NO2 <- 46
mgm3_to_ppm(mgm3 = 1, mw = 46, temp = 293, press = 1013)
}
