% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ppm_ppb.R
\name{ppb_to_ugm3}
\alias{ppb_to_ugm3}
\title{ppb to µg/m3}
\usage{
ppb_to_ugm3(ppb, mw, temp = 293.15, press = 1013)
}
\arguments{
\item{ppb}{numeric vector of volume concentrations in parts per billion (10^-9)}

\item{mw}{molecular weight (g/mol)}

\item{temp}{temperature in K (default = 293 K)}

\item{press}{pressure in hPa (default = 1013 hPa)}
}
\value{
numeric vector of mass concentrations
}
\description{
convert from ppb to µg/m3
}
\examples{
NO2 <- 46
ppb_to_ugm3(ppb = 100, mw = 46, temp = 293, press = 1013)
}
