% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_to_mgm3_and_reverse.R
\name{ppm_to_mgm3}
\alias{ppm_to_mgm3}
\title{ppm to mg/m3}
\usage{
ppm_to_mgm3(ppm, mw, temp = 293.15, press = 1013)
}
\arguments{
\item{ppm}{numeric vector of volume concentrations in parts per billion (10^-6)}

\item{mw}{molecular weight (g/mol)}

\item{temp}{temperature in K (default = 293 K)}

\item{press}{pressure in hPa (default = 1013 hPa)}
}
\value{
numeric vector of mass concentrations
}
\description{
convert from ppm to mg/m3
}
\examples{
NO2 = 46
ppm_to_mgm3(ppm = 100,  mw = 46, temp = 293, press = 1013)
}
