% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfun-package.R
\docType{package}
\name{rhfun-package}
\alias{rhfun}
\alias{rhfun-package}
\title{rhfun: R handy functions: the new collection}
\description{
A new collection of my handy R functions. The reboot of an old package of mine.
}
\author{
\strong{Maintainer}: massimo bressan \email{massimo.bressan@arpa.veneto.it}

}
\keyword{internal}
