% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{temp_from_to}
\alias{temp_from_to}
\title{Temperature conversion utility}
\usage{
temp_from_to(temp, from = "C", to = "K")
}
\arguments{
\item{temp}{numeric vector of temperature}

\item{from}{character scale "from"}

\item{to}{character scale "to}
}
\value{
vector of converted temperature
}
\description{
Convert between Celsius, Fahrenheit, and Kelvin seamlessly (hopefully!)
}
\examples{
temp_from_to(100, "C", "F")
temp_from_to(212, "F", "C")
temp_from_to(300, "K", "C")
temp_from_to(100, "C", "K")
temp_from_to(98.6, "F", "K")
temp_from_to(273.15, "K", "F")
temp_from_to(0, "C", "C")

}
