% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppb_to_ugm3_and_reverse.R
\name{ugm3_to_ppb}
\alias{ugm3_to_ppb}
\title{µg/m3 to ppb}
\usage{
ugm3_to_ppb(ugm3, mw, temp = 293.15, press = 1013)
}
\arguments{
\item{ugm3}{numeric vector of mass concentrations in ug per cubic meter (m^-3)}

\item{mw}{molecular weight (g/mol)}

\item{temp}{temperature in K (default = 293 K)}

\item{press}{pressure in hPa (default = 1013 hPa)}
}
\value{
numeric vector of volume concentrations in parts per billion (10^-9)
}
\description{
convert from µg/m3 to ppb
}
\examples{
NO2 = 46
ugm3_to_ppb(ugm3 = 1,  mw = 46, temp = 293, press = 1013)
}
