% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wind.R
\name{uv_to_wdws}
\alias{uv_to_wdws}
\title{convert u and v components to wind direction and wind speed}
\usage{
uv_to_wdws(u, v)
}
\arguments{
\item{u}{a vector of u components}

\item{v}{a vector of v components}
}
\value{
a matrix
}
\description{
this function converts u and v components to wind direction and wind speed
}
\examples{
set.seed(123)
wd <- as.integer(rnorm(10, 180, 90))

set.seed(123)
ws <- rnorm(10, 4, 1)

## convert to u and v
wdws_to_uv(wd, ws)

## convert back
uv <- wdws_to_uv(wd, ws)
uv_to_wdws(uv[, 1], uv[, 2])

}
\seealso{
\code{\link{wdws_to_uv}}
}
\author{
Credits to Tim Appelhans
https://raw.githubusercontent.com/environmentalinformatics-marburg/Rsenal/master/R/uv2wdws.R
}
