% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wind.R
\name{wdws_to_uv}
\alias{wdws_to_uv}
\title{convert wind direction and wind speed to u and v components}
\usage{
wdws_to_uv(wd, ws)
}
\arguments{
\item{wd}{a vector of wind directions}

\item{ws}{a vector of wind speeds}
}
\value{
a matrix
}
\description{
this function converts wind direction and wind speed to u and v components

Some specifications about the function code
Why in the code function there is a negative sign in ws?
Because of Wind Direction Convention:
Wind direction (wd) is given as the direction from which the wind is coming, measured in degrees clockwise from North (0°).
U and V components represent the wind vector pointing in the direction the wind is blowing.
Wind direction is measured from where the wind originates, but sine and cosine functions assume an angle that indicates where the wind is going.
Without the negative sign, the computed vector would point in the opposite direction, which is incorrect.
}
\examples{
set.seed(123)
wd <- as.integer(rnorm(10, 180, 90))

set.seed(123)
ws <- rnorm(10, 4, 1)

## convert to u and v
wdws_to_uv(wd, ws)

## convert back
uv <- wdws_to_uv(wd, ws)
uv_to_wdws(uv[, 1], uv[, 2])

}
\seealso{
\code{\link{uv_to_wdws}}
}
\author{
credits to Tim Appelhans
https://raw.githubusercontent.com/environmentalinformatics-marburg/Rsenal/master/R/wdws2uv.R
}
