% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wind.R
\name{wind_sect_to_deg}
\alias{wind_sect_to_deg}
\title{Convert wind sectors to degrees}
\usage{
wind_sect_to_deg(wind_sect)
}
\arguments{
\item{wind_sect}{character vector of wind sectors}
}
\value{
numeric vector of wind degrees
}
\description{
Converts wind sectors from North to corresponding wind degrees.
By default are considered 16 wind sectors with each amplitude of 22.5 degrees.
}
\examples{
wind_sect <- c("N","NNE","NE", "E", "ESE", "SE", "SSE", "S", "W", "WNW", "NW", "NNW")
wind_sect_to_deg(wind_sect)

}
