% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_convert.R
\name{wind_uv_to_ds}
\alias{wind_uv_to_ds}
\title{Convert u and v components to wind direction and wind speed}
\usage{
wind_uv_to_ds(u, v)
}
\arguments{
\item{u}{a vector of u components}

\item{v}{a vector of v components}
}
\value{
a matrix
}
\description{
This function converts u and v components to wind direction and wind speed
}
\examples{
set.seed(123)
d <- as.integer(rnorm(10, 180, 90))
set.seed(123)
s <- rnorm(10, 4, 1)
## convert to u and v
wind_ds_to_uv(d, s)
## convert back
uv <- wind_ds_to_uv(d, s)
wind_uv_to_ds(uv[, 1], uv[, 2])

}
\seealso{
\code{\link{wind_ds_to_uv}}
}
\author{
Credits to Tim Appelhans
https://raw.githubusercontent.com/environmentalinformatics-marburg/Rsenal/master/R/uv2wdws.R
}
