% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_yday.R
\name{yday_to_date}
\alias{yday_to_date}
\title{From year-day to date}
\usage{
yday_to_date(yday)
}
\arguments{
\item{yday}{string  in the form "YYYY-YDAY" representing input year-day (a.k.a. julian-day)}
}
\value{
string in the form "YYY-MM-DD"representing output date
}
\description{
Convert year-day (a.k.a. julian-day) in the form "YYYY-YDAY" to date in the form "YYY-MM-DD"
}
\examples{
yday_to_date("2022-150")
}
