% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_angle.R
\name{deg_to_car}
\alias{deg_to_car}
\title{Angle degrees to Cartesian x, y coordinates}
\usage{
deg_to_car(deg, r = 1)
}
\arguments{
\item{deg}{a numeric vector of angle degrees}

\item{r}{numeric, ray of the unitary circle}
}
\value{
a tibble with degrees, phi in radians, x and y
}
\description{
Convert angle degrees to cartesian coordinates x and y.
It is internally also taking care of the conversion from input angles as degrees to radians
}
\examples{
deg <- c(0, 30, 45, 60, 90, 180, 270, 360)
deg_to_car(deg)
deg_to_car(0)      # Expected: (1, 0)
deg_to_car(90)     # Expected: (0, 1)
deg_to_car(180)    # Expected: (-1, 0)
deg_to_car(270)    # Expected: (0, -1)
deg_to_car(45, 2)  # Expected: (√2, √2)
}
