% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_angle.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Angle degrees to radians}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{a numeric vector of angles degrees}
}
\value{
a numeric vector of corresponding radians
}
\description{
Convert angles degrees to radians
}
\examples{
deg <- c(0, 30, 45, 60, 90, 180, 270, 360)
deg_to_rad(deg)
}
