% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wind.R
\name{wind_deg_to_sect}
\alias{wind_deg_to_sect}
\title{wind_deg_to_sect}
\usage{
wind_deg_to_sect(wind_deg)
}
\arguments{
\item{wind_deg}{numeric vector of wind degrees}
}
\value{
character vector of wind sectors
}
\description{
convert wind degrees to sectors
}
\examples{
wind_deg <- c(0, seq(11.25, (360-11.25), by = 22.5), 360)
wind_deg_to_sect(wind_deg)

}
