% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wind.R
\name{wind_sect_to_deg}
\alias{wind_sect_to_deg}
\title{wind_sect_to_deg}
\usage{
wind_sect_to_deg(wind_sect)
}
\arguments{
\item{wind_sect}{character vector of wind sectors}
}
\value{
numeric vector of wind degrees
}
\description{
convert wind sectors to degrees
}
\examples{
wind_sect <- c("N","NNE","NE", "E", "ESE", "SE", "SSE", "S", "W", "WNW", "NW", "NNW")
wind_sect_to_deg(wind_sect)

}
